+++ examples/gltf_loader/run.sh
@@ -1,0 +1,3 @@
+#!/bin/sh
+
+bundle exec opal -qopal/js_wrap/three -Rserver example.rb