+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## [Unreleased]
 
+- Not specified here improvements to JSWrap
+- 3DObject#<< as an alias to #add
+- Full support for examples (ie. loaders)
+- New examples: clipping_intersection, gltf_loader
+
 ## [0.1.0] - 2021-12-14
 
 - Initial release